Undid some stupid int->int16_t conversions.

+++ drizzled/log_event.cc
@@ -8229,7 +8229,7 @@ Heartbeat_log_event::Heartbeat_log_event(const char* buf, uint event_len,
 st_print_event_info::st_print_event_info()
   :flags2_inited(0), sql_mode_inited(0),
    auto_increment_increment(0),auto_increment_offset(0), charset_inited(0),
-   lc_time_names_number(UINT16_MAX),
+   lc_time_names_number(UINT32_MAX),
    charset_database_number(ILLEGAL_CHARSET_INFO_NUMBER),
    thread_id(0), thread_id_printed(false),
    base64_output_mode(BASE64_OUTPUT_UNSPEC), printed_fd_event(false)
